+++ apps/native-component-list/src/screens/UI/UIScreen.android.tsx
@@ -115,6 +115,14 @@ export const UIScreens = [
     },
   },
   {
+    name: 'SyncSwitch component',
+    route: 'ui/sync-switch',
+    options: {},
+    getComponent() {
+      return optionalRequire(() => require('./SyncSwitchScreen'));
+    },
+  },
+  {
     name: 'Shape component',
     route: 'ui/shape',
     options: {},